// src/lib/cards/xmage/tokens.generated.ts
export const XMAGE_TOKENS: Record<string, TokenEntry> = {
// ...
 "SoldierToken": {
  "spec": {
   "name": "Soldier",
   "typeLine": "Creature — Soldier",
   "power": "1",
   "toughness": "1",
   "colorIdentity": [
    "W"
   ]
  },
  "otherAbilities": []
 },
// ...
 "TreasureToken": {
  "spec": {
   "name": "Treasure",
   "typeLine": "Artifact — Treasure"
  },
  "otherAbilities": [
   "TreasureAbility"
  ]
 },
// ...
};